chore: 데이터 포멧팅 유틸 분리
