+++ src/shared/lib/format/index.ts
@@ -1,0 +1,1 @@
+export * from './format';
