+++ src/shared/lib/index.ts
@@ -1,7 +1,4 @@
 export { createFormData } from './formDataUtil';
-export { formatDateInput } from './inputFormatters';
-export { getTimeDifference, formatDateToYyyyMmDd } from './format';
-export { formatPhoneNumber } from './formatters';
 export * from './tokenManager';
 export * from './notificationPermission';
 export * from './axiosErrorParser';
@@ -12,3 +9,4 @@ export { getClientFCMToken } from './fcmToken';
 export * from './capacitor';
 export * from './analytics';
 export * from './sanitizer';
+export * from './format';
